uint16 was too small to hold some multiplication products; uint16 was replaced with uint64; program now achieves expected result with puzzle input.

+++ day11/puzzle1/main.cpp
@@ -10,7 +10,7 @@ void clearOutput(std::string filename) {
     output.close();
 }//clearOutputFile
 
-void writeOutput(std::string filename, std::vector<uint16_t>* values, int blinks) {
+void writeOutput(std::string filename, std::vector<uint64_t>* values, int blinks) {
     std::ofstream output;
     output.open(filename, std::ios::app);
     output << blinks << ": ";
